fixing

+++ packages/carrel/carrel_db/entity/src/implementation/archive_traits.rs
@@ -27,7 +27,7 @@ impl From<AddArchiveDto> for archive::ActiveModel {
 impl ArchiveTrait for archive::Model {
     fn into_archive(self) -> Archive {
         Archive {
-            id: self.id as i64,
+            id: self.id,
             uuid: self.uuid,
             project_uuids: vec![],
             name: self.name,
